requirements.txt
streamlit
spotipy
pandas
numpy
matplotlib
wordcloud
scikit-learn
python-dotenv
